ft(validations):Restructuring validation files and message
- moved the validation files from the middleware into the validation folder
- mapped the errors from the schemas so that they can be displayed all in an array
