+++ src/utils/schemas.js
@@ -1,19 +1,28 @@
 import Joi from '@hapi/joi';
 
-const signupSchema = {
+const signupSchema = Joi.object().keys({
   firstName: Joi.string().strict().trim().required(),
   lastName: Joi.string().strict().trim().required(),
   email: Joi.string().strict().trim().email()
     .required(),
   password: Joi.string().alphanum().min(8).required()
-    .strict(),
-};
+}).options({
+  abortEarly: false,
+  language: {
+    key: '{{key}} '
+  }
+});
 
-const signinSchema = {
+const signinSchema = Joi.object().keys({
   email: Joi.string().strict().trim().email()
     .required(),
-  password: Joi.string().alphanum().min(8).required(),
-};
+  password: Joi.required()
+}).options({
+  abortEarly: false,
+  language: {
+    key: '{{key}} '
+  }
+});
 
 export default {
   '/auth/signup': signupSchema,
